+++ yash-quote/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to `yash-quote` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - Unreleased
+
+### Added
+
+- The `Quoted` struct
+    - The `as_raw` and `needs_quoting` methods
+    - `impl std::fmt::Display for Quoted<'_>`
+    - `impl<'a> From<&'a str> for Quoted<'a>`
+    - `impl<'a> From<Quoted<'a>> for Cow<'a, str>`
+- The `quoted` function
+
 ## [1.0.1] - 2023-01-05
 
 ### Fixed